fix(gui): stop in-process server on app teardown (Android lifecycle)
An Android back/exit only finishes the Activity while the process stays
alive — the in-process Rust WebDAV server (and its scan/unlock cache in
static state) survived into the next session. Consequences:
  - a fresh launch reused the stale accepted keys, so ANY password
    (including wrong/empty) unlocked every file and skipped verification
  - each start stacked a second server on the same files (the old
    shutdown flag was overwritten), leaking server threads + file
    handles and eventually stalling verification forever

Fixes:
  - app.dart: AppLifecycleState.detached now stops the server, clearing
    all Rust state so the next session re-scans and re-verifies (wrong
    passwords rejected again). Backgrounding (paused) is unaffected.
  - server.rs: start() shuts down any leftover server from a dead
    session and resets the scan/unlock cache as defense-in-depth
    (multi-password incremental flow keeps its cache untouched)
  - qr_gen.rs: fix clippy unnecessary_cast warnings

+++ crates/rcrm-flutter-bridge/src/qr_gen.rs
@@ -111,7 +111,7 @@ mod tests {
 		assert_eq!(w, QR_SIZE);
 		assert_eq!(h, QR_SIZE);
 		let px = |x: usize, y: usize| {
-			let i = (y as usize * QR_SIZE as usize + x as usize) * 4;
+			let i = (y * QR_SIZE as usize + x) * 4;
 			rgba[i]
 		};
 		// Four corners are quiet zone → pixel-exact white (255) in lossless.
